+++ scripts/devloop.sh
@@ -1,5 +1,5 @@
-while inotifywait **/*.rb Gemfile unparser.gemspec; do
+while inotifywait ../mutant/**/*.rb **/*.rb Gemfile unparser.gemspec; do
   bundle exec rspec spec/unit -fd --fail-fast --order default \
-    && bundle exec ./mutant.sh --since master -- 'Unparser*' \
+    && bundle exec mutant run --zombie --since master --fail-fast -- 'Unparser*' \
     && bundle exec rubocop
 done